requirements.txt
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref @ file:///croot/asgiref_1724072470684/work
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1734473278428/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
clip==0.2.0
conda @ file:///croot/conda_1736277811838/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
cryptography @ file:///croot/cryptography_1724940562255/work
distro @ file:///croot/distro_1714488253808/work
dj-database-url==2.3.0
Django==5.1.4
django-filter==25.1
django-on-heroku==1.1.2
djangorestframework==3.15.2
frozendict @ file:///croot/frozendict_1713194832637/work
idna @ file:///croot/idna_1714398848350/work
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1725563367192/work/libmambapy
Markdown==3.7
menuinst @ file:///croot/menuinst_1723567589013/work
packaging @ file:///croot/packaging_1720101850331/work
pillow==11.1.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
psycopg2-binary==2.9.10
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
requests @ file:///croot/requests_1721410876868/work
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
setuptools==75.1.0
sqlparse==0.5.3
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1727769808118/work
wheel==0.44.0
whitenoise==6.9.0
wsl==0.12
zstandard @ file:///croot/zstandard_1728569189425/work
